dvc486: accessibilty fixes
dvc486: add slider styles

dvc486: add updated sliders

dvc486: overwrite old visualisation
